refactor(discord-adapter.c): split to discord-adapter_refcount.c
