+++ Makefile
@@ -33,6 +33,7 @@ THIRDP_OBJS   = $(THIRDP_DIR)/sha1.o           \
 DISCORD_OBJS  = $(SRC_DIR)/concord-once.o              \
                 $(SRC_DIR)/discord-adapter.o           \
                 $(SRC_DIR)/discord-adapter_ratelimit.o \
+                $(SRC_DIR)/discord-adapter_refcount.o  \
                 $(SRC_DIR)/discord-client.o            \
                 $(SRC_DIR)/discord-gateway.o           \
                 $(SRC_DIR)/discord-timer.o             \
